+++ src/app/providers/translator.service.ts
@@ -1,4 +1,3 @@
-
 import { Injectable } from '@angular/core';
 // import * as translations from '../utils/translations';
 import { TranslateService } from '@ngx-translate/core';