+++ src/chat21-core/utils/constants.ts
@@ -81,3 +81,6 @@ export const URL_SOUND = 'assets/pling.mp3';
 // PARAMETERS
 export const PLATFORM_MOBILE = 'mobile';
 export const PLATFORM_DESKTOP = 'desktop';
+
+// STORAGE
+export const STORAGE_PREFIX = 'tiledesk_widget_';